# swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "OpenAPI definition",
        "version": "v0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
    "paths": {
        "/ws/{organization}/{project}/_apis/pipelines/pipelinepermissions/{resourceType}/{resourceId}": {
            "post": {
                "tags": [
                    "pipelene-permission-cli"
                ],
                "operationId": "getPipelinePermissions",
                "parameters": [
                    {
                        "name": "organization",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "project",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "resourceType",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "resourceId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ResourcePipelinePermissions"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "pipelene-permission-cli"
                ],
                "operationId": "updatePipelinePermissions",
                "parameters": [
                    {
                        "name": "organization",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "project",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "resourceType",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "resourceId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ResourcePipelinePermissions"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "IdentityRef": {
                "type": "object"
            },
            "Permission": {
                "type": "object",
                "properties": {
                    "authorized": {
                        "type": "boolean"
                    },
                    "authorizedBy": {
                        "$ref": "#/components/schemas/IdentityRef"
                    },
                    "authorizedOn": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "PipelinePermission": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    }
                }
            },
            "Resource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    }
                }
            },
            "ResourcePipelinePermissions": {
                "type": "object",
                "properties": {
                    "allPipelines": {
                        "$ref": "#/components/schemas/Permission"
                    },
                    "pipelines": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PipelinePermission"
                        }
                    },
                    "resource": {
                        "$ref": "#/components/schemas/Resource"
                    }
                }
            }
        }
    }
}